refactor(ios): split composer model selection
